Finish CV screen. Problems getting it to animate
Long time since last commit. I've pretty much finished the CV page. We
have a js file exporting entries for the card which are rendered with
the perspective effect. I don't really know how to do css animations on
this stuff and it's actually really annoying. All that's left now is to
do the contact me page which should be dead easy